fix angle comparison

+++ src/main/java/ch/njol/skript/expressions/ExprVectorAngleBetween.java
@@ -62,7 +62,7 @@ public class ExprVectorAngleBetween extends SimpleExpression<Number> {
 		Vector second = this.second.getSingle(event);
 		if (first == null || second == null)
 			return null;
-		return CollectionUtils.array(first.angle(second) * (float) 180 / Math.PI); // rad to deg
+		return CollectionUtils.array(first.angle(second) * (180.0 / Math.PI)); // rad to deg
 	}
 
 	@Override
